requirements.txt
requests
pillow
pyyaml
pytest
python-dotenv
